fix: reorganize cli into its own directory
